requirements.txt
pytest
selenium
webdriver-manager
allure-pytest
python-dotenv
black
flake8
pytest-xdist